M2-5088: Move settings modal into Popups, add tests for Actions
