+++ src/shared/components/NavigationMenu/Actions/Actions.const.ts
@@ -1,0 +1,1 @@
+export const DATA_TESTID_ACTIONS_EXPORT_DATA = 'builder-applet-settings-export-data';
