+++ src/shared/features/AppletSettings/ExportDataSetting/ExportDataSetting.const.ts
@@ -1,0 +1,2 @@
+export const DATA_TESTID_EXPORT_DATA_SETTINGS_POPUP = 'applet-settings-export-data-settings-popup';
+export const DATA_TESTID_EXPORT_DATA_EXPORT_POPUP = 'applet-settings-export-data-export-popup';
