+++ src/shared/features/AppletSettings/ExportDataSetting/Popups/ExportSettingsPopup/ExportSettingsPopup.types.ts
@@ -1,0 +1,5 @@
+export type ExportSettingsPopupProps = {
+  isOpen: boolean;
+  onClose: () => void;
+  onExport: () => void;
+};
