+++ src/shared/features/AppletSettings/ExportDataSetting/Popups/ExportSettingsPopup/index.ts
@@ -1,0 +1,1 @@
+export * from './ExportSettingsPopup';
